Why We Ask Candidates ... Queries/Comments JamesRossPR
What do we mean by "Repo"? Will readers know what that means? Should we spell out?

Is "Git" the same as "Github"? It seems like you're talking about two different things--but that Git might also be a sort of nickname for GitHub. Can we make the distinction clear? Just a few words that briefly inform the reader what they are and that they are two different things. 

I broke up the first paragraph into two. We need a bit of detail in the opening paragraph. Just a few words about what is so new about Git, why it is transformative, what paradigms are shifting. Why is it so great? Just a bit, because we're

What do you mean by "submit to Git"? Does this mean simply to use it? Or "to submit work through Git"?

The paragraphs that begin with "Linus Torvalds" and "Unfortunately" seem to wander off topic. Perhaps delete them and begin the following paragraph with: Another reason we want candidates to work within Git is that it gives us a closer look at our programmers (or coders or whatever their job is).

In general, is it OK to use first person in these articles? "I think ..."

If you keep the paragraph about how Git is not well understood: "Even in technology circles private ownership still dominates." Explain how this keeps people from using or understanding Git?

Rewrote some of the paragraph starting at 19. OK?

"It they stoned the challenge we provided them, or if they go that extra mile." I don't know what this means. Transcription error?

"So, providing answers on Git," Answers to what? Does this mean: "working on Git" "Using Git" Clarify?

We're getting some details about the benefits of Git in these last few paragraphs. But are these things really unique to Git?

I think if you delete the two paragraphs that I suggest, you have a nicely structured article here. Maybe you can take the paragraph about why Git is not widely used or well-understood and spin that5 into a separate article.