Importance of Learning Comments/Queries JamesRossPR
That word "tension" seems funny. I get it, but I wonder if another word might be more meaningful here: obstacles, distractions, digression, something like that. I'm not going to change the word, but think about it.

Same with word "irrational." Again, I get it but ... Maybe naive, reckless ... What do you think?

"because that what allows you to have the priors" Not quite sure of the meaning of this. Maybe: "that is your foundation"

Perhaps delete: "In a way, having a certain level of irrationality is actually very important, because that's what allows you to start down that path." from paragraph starting at 17.

I think this is a fairly strong article. Solid advice.