Elon Musk Programmer  Comments/Queries JamesRossPR
"in a way the whole life-cycle process itself is the machine." I'm not quite sure what this means. Clarify? This is a good place for a clear, concrete statement.

"They use software, in a way, that outstrips the attempts by other companies." I'm a little confused by this. Do you mean the software is much better than that used by other companies? If so, delete ",in a way," Or do you mean that they use software IN A WAY that other companies don't? Delete commas. Also, are you saying they EMPLOY software in their products in a way that other companies don't? Finally, is the fact that the Tesla cars update themselves an example of this? Should rewrite this sentence. 

Pretty solid article. Just clear up the above issues.