Don't Give Problems Queries/Comments JamesRossPR
"the team members work on faith-baed data." Should this be "faith-based data"?

"It makes excellent business sense to handle project tasks in-house." Last paragraph: Maybe to better connect this graph to the rest of the piece we can begin like this: Finally, one of the best and most practical ways to keep from giving others your problems is to handle project tasks in-house. What do you think?

Nice article. Maybe rethink the ending. Starting with "If you look at project management..." the ideas seem to wander away from the main idea. It says, basically, "we should be doing good things." Everything up to that point contributes strongly to the topic. Perhaps craft an ending that drives home the importance of solving your own problems,